Set per-worker test session names
